requirements.txt
fastai
tensorflow 
pandas
Pillow
streamlit
numpy
clip
matplotlib
sentence-transformers
torch
torchvision
pretrainedmodels
tqdm